requirements.txt
fastapi
uvicorn
python-dotenv
requests
aiohttp
pydantic
pydantic-settings
litellm
tavily-python
colorama
streamlit